[hive] Use default hive-site.xml if hive-conf-dir is not set (#1024)
