+++ paimon-hive/paimon-hive-catalog/src/test/java/org/apache/paimon/hive/HiveCatalogTest.java
@@ -83,4 +83,35 @@ public class HiveCatalogTest extends CatalogTestBase {
                 .hasRootCauseInstanceOf(IllegalStateException.class)
                 .hasRootCauseMessage("Table name[NEW_TABLE] cannot contain upper case");
     }
+
+    private static final String HADOOP_CONF_DIR =
+            Thread.currentThread().getContextClassLoader().getResource("hadoop-conf-dir").getPath();
+
+    private static final String HIVE_CONF_DIR =
+            Thread.currentThread().getContextClassLoader().getResource("hive-conf-dir").getPath();
+
+    @Test
+    public void testHadoopConfDir() {
+        HiveConf hiveConf = HiveCatalog.createHiveConf(null, HADOOP_CONF_DIR);
+        assertThat(hiveConf.get("fs.defaultFS")).isEqualTo("dummy-fs");
+    }
+
+    @Test
+    public void testHiveConfDir() {
+        try {
+            testHiveConfDirImpl();
+        } finally {
+            cleanUpHiveConfDir();
+        }
+    }
+
+    private void testHiveConfDirImpl() {
+        HiveConf hiveConf = HiveCatalog.createHiveConf(HIVE_CONF_DIR, null);
+        assertThat(hiveConf.get("hive.metastore.uris")).isEqualTo("dummy-hms");
+    }
+
+    private void cleanUpHiveConfDir() {
+        // reset back to default value
+        HiveConf.setHiveSiteLocation(HiveConf.class.getClassLoader().getResource("hive-site.xml"));
+    }
 }
