+++ paimon-hive/paimon-hive-catalog/src/test/resources/hadoop-conf-dir/core-site.xml
