+++ paimon-hive/paimon-hive-catalog/src/test/resources/hive-conf-dir/hive-site.xml
